FDG-6274 refactor section
