+++ src/layouts/explainer/sections/national-debt/funding-programs-and-services/funding-programs-and-services.spec.js
@@ -1,0 +1,50 @@
+import Analytics from '../../../../../utils/analytics/analytics';
+import { render } from '@testing-library/react';
+import React from 'react';
+import FundingProgramsAndServices from './funding-programs-and-services';
+
+describe('Funding Programs & Services', () => {
+  it('calls the appropriate analytics event when links are clicked on', () => {
+    const spy = jest.spyOn(Analytics, 'event');
+    const { getByText, getAllByText } = render(<FundingProgramsAndServices />);
+
+
+    const accordion = getByText('What are some of the major spending categories?');
+    accordion.click();
+    const usaSpending = getAllByText('USAspending.gov');
+    const objectClass = getByText('Object Class');
+    const budgetFunction = getByText('Budget Function');
+
+    usaSpending[1].click();
+    expect(spy).toHaveBeenCalledWith({
+      category: 'Explainers',
+      action: `Citation Click`,
+      label: 'Debt - Funding Programs & Services'
+    });
+    spy.mockClear();
+
+    usaSpending[0].click();
+    expect(spy).toHaveBeenCalledWith({
+      category: 'Explainers',
+      action: `Citation Click`,
+      label: 'Debt - What are the major spending categories?'
+    });
+    spy.mockClear();
+
+    objectClass.click();
+    expect(spy).toHaveBeenCalledWith({
+      category: 'Explainers',
+      action: `Citation Click`,
+      label: 'Debt - What are the major spending categories?'
+    });
+    spy.mockClear();
+
+    budgetFunction.click();
+    expect(spy).toHaveBeenCalledWith({
+      category: 'Explainers',
+      action: `Citation Click`,
+      label: 'Debt - What are the major spending categories?'
+    });
+    spy.mockClear();
+  });
+});
