+++ src/layouts/explainer/sections/national-debt/funding-programs-and-services/spending-categories-accordion/spending-categories-accordion.module.scss.d.ts
@@ -63,30 +63,10 @@ export const paleBg: string;
 export const debtExplainerPrimary: string;
 export const debtExplainerSecondary: string;
 export const debtExplainerLightSecondary: string;
-export const keyTakeawaysContent: string;
-export const noMarginBottom: string;
-export const iconBackground: string;
+export const fundingProgramAccordion: string;
+export const spendingCategoriesAccordionContent: string;
+export const spendingCategoriesTable: string;
+export const row: string;
+export const firstColumn: string;
+export const secondColumn: string;
 export const icon: string;
-export const offsetIcon: string;
-export const postGraphAccordionContainer: string;
-export const DebtOverLast100yContainer: string;
-export const growingNationalDebtSectionGraphContainer: string;
-export const debtTrendsOverTimeSectionGraphContainer: string;
-export const title: string;
-export const simple: string;
-export const titleBreakdown: string;
-export const subTitle: string;
-export const headerContainer: string;
-export const header: string;
-export const subHeader: string;
-export const growingNationalDebtSectionGraph: string;
-export const footerContainer: string;
-export const lineChartContainer: string;
-export const postGraphContent: string;
-export const diveDeeperQuoteRight: string;
-export const diveDeeperQuote: string;
-export const diveDeeperQuoteLeft: string;
-export const diveDeeperCitation: string;
-export const diveDeeperLink: string;
-export const debtAccordion: string;
-export const debtCeilingAccordion: string;
