+++ src/layouts/explainer/sections/national-debt/national-debt.spec.js
@@ -43,53 +43,6 @@ describe('National Debt explainer page sections', () => {
   });
 });
 
-
-describe('Funding Programs & Services', () => {
-  it('calls the appropriate analytics event when links are clicked on', () => {
-    const spy = jest.spyOn(Analytics, 'event');
-    const { getByText, getAllByText } = render(<FundingProgramsSection />);
-
-
-    const accordion = getByText('What are some of the major spending categories?');
-    accordion.click();
-    const usaSpending = getAllByText('USAspending.gov');
-    const objectClass = getByText('Object Class');
-    const budgetFunction = getByText('Budget Function');
-
-    usaSpending[1].click();
-    expect(spy).toHaveBeenCalledWith({
-      category: 'Explainers',
-      action: `Citation Click`,
-      label: 'Debt - Funding Programs & Services'
-    });
-    spy.mockClear();
-
-    usaSpending[0].click();
-    expect(spy).toHaveBeenCalledWith({
-      category: 'Explainers',
-      action: `Citation Click`,
-      label: 'Debt - What are the major spending categories?'
-    });
-    spy.mockClear();
-
-    objectClass.click();
-    expect(spy).toHaveBeenCalledWith({
-      category: 'Explainers',
-      action: `Citation Click`,
-      label: 'Debt - What are the major spending categories?'
-    });
-    spy.mockClear();
-
-    budgetFunction.click();
-    expect(spy).toHaveBeenCalledWith({
-      category: 'Explainers',
-      action: `Citation Click`,
-      label: 'Debt - What are the major spending categories?'
-    });
-    spy.mockClear();
-  });
-});
-
 jest.mock('../../../../hooks/useBeaGDP', () => {
   return () => mockBeaGDPData;
 });
