+++ src/pages/americas-finance-guide/afg-overview.module.scss.d.ts
@@ -75,12 +75,6 @@ export const iconBackground: string;
 export const icon: string;
 export const offsetIcon: string;
 export const postGraphAccordionContainer: string;
-export const fundingProgramAccordion: string;
-export const spendingCategoriesAccordionContent: string;
-export const spendingCategoriesTable: string;
-export const row: string;
-export const firstColumn: string;
-export const secondColumn: string;
 export const DebtOverLast100yContainer: string;
 export const growingNationalDebtSectionGraphContainer: string;
 export const debtTrendsOverTimeSectionGraphContainer: string;
